Tmp hack: Override `opt-level` when compiling libcore
__aeabi_* compiler intrinsics are the bane of my existence...

For some reason `opt-level=2` includes an `__aeabi_memset` (an GCC
intrinsic available during linking) when compiling libcore that doesn't
get resolved when compiling main.o, even though it seems to be available
for some other crates. This works for now (i.e. compiles from scratch,
and the code functions as expected, with blinking and everything).
However, should be fixed ASAP, this is obviously a stupid bug.