More progress on enabling interrupts for console reading.
This change in progress highlights the need for the following:
  1) Device/Platform driver interrupt handling infrastructure.
  2) Robust kernel-to-user calling/scheduling mechanisms.

As it stands, device or platform drivers cannot request to be notified when an
interrupt is received by the kernel. Further, it's unclear who's responsibility
it is to enable certain interrupts. Here's the proposition:

On interrupts: At compile time, via the platform tree, platform drivers should
be able to request lines on the NVIC. For example: USART(NVICINT@27). This
guarantees that only one platform driver can listen to an interrupt request at a
time. The type passed in to the driver should allow for enabling the NVIC
interrupt and the registering of a callback for said interrupt. This way, the
kernel knows which interrupts are enabled when, can catch those, and can
schedule them to be sent to the proper driver when needed.

On calling user level code: All calls should be done through events as described
in the event-driven infrastructure document by @SergioBenitez. When some app
or driver publishes information, the kernel will take care of waking up the
processes that need to take a look at their event logs.