Peripheral function selection can't be in drivers
Peripheral function is platform specific (sam4l) *and* pin specific. You
have to have global knowledge of the platform wiring to know which pins
to use and how to select them. For exmple, the console can't just select
function A, because sometimes A means USART, while for other pins
another function is USART. Same issue with putting them in platform drivers.

So logical place is to select pin functions as the platform driver is
being instantiated but outside of both the plaform and device driver.