Parameter-less device drivers now working.
Device drivers initialized without parameters in the device tree are now parsed
correctly. Unfortunately, this exercise resulted in the discovery of an issue:
since syntax extensions are constrained to the AST of the extension itself, they
cannot refer to variables declared outside the syntax extension's block. This
is an issue as we hope to use variables declared through the platform_tree!
driver in the device_tree! driver.

Ideally, we'd be able to somehow point the syntax extension's AST to its parent
so that all platform_tree! variables are present. This requires further
research.