Using slices instead of arrays is just better
The bug mentioned in a previous commit seems like it *could* be a
compiler error that occurs when copying an array from a static variable
onto the stack under certain circumstances (still trying to pinpoint).
Why copy instead of just reference? Well, yeah, that was a mistake.
Changing to a reference side-steps this problem entirely, so two birds
with one stone.